Class Today_Week6
In this it the changes for today.
It includes:
The code for the bobber behavior
Reorganized / moved around / created folders:

Note: Monica, I moved your solo edits of the menu code to "_soloWork/Monica"
Also Note: Lona, you might have to resubmit your boat model. For some reason I didn't see it in the recent updates

Note to everyone:
PLEASE BE ORGANIZED!!!!